Move the SupernodalLU sweep proof to a QA-only file
`test/runtests.jl` includes `qa/allocations.jl` from two groups, and they do not
cover the same Julia versions:

    [QA]              versions = ["lts", "1"]
    [AppleAccelerate] versions = ["lts", "1", "pre"]

So AllocCheck ran on `pre` through a group that was never meant to carry it.

The file mixes two things with different portability. The direct-BLAS
refactorization proofs hold on every release - that contract is why the
AppleAccelerate group exists. The SupernodalLU sweep proof does not: above
`PANEL_BLAS_CUTOFF` the sweeps hand off to LinearAlgebra's `ldiv!`/`mul!`,
whose wrappers leave a `generic_trimatdiv!` dynamic dispatch plus boxed
`MulAddMul`/`SubArray` values that AllocCheck reports on 1.10 and 1.13. Only
the second one belongs behind QA's version set, so only it moves.

Measured, same two matrices as the testset:

    Julia 1.10.11  x86_64-Linux   18 static findings   runtime 0 B / 0 B
    Julia 1.12.6   x86_64-Linux    0 static findings   runtime 0 B / 0 B

18 findings on 1.10 x86_64-Linux is the same count #1119 measured on 1.10
aarch64-Darwin, and 1.12 is clean on both, so this is the stdlib and the Julia
version - not the platform, and not this package's code. Worth noting the
coverage is narrower than the toml suggests: only one QA job materialises
(`QA (julia 1)`), so the static proof runs on 1.12 alone.

The static proof keeps a version guard, since that is a limit of the prover.
`RUNTIME_MULTIRHS_ZERO` is dropped: it gated a *runtime* allocation assertion,
and on 1.11 that assertion fails because the multi-RHS path really does
allocate (6144 B, per #1119's table). Skipping it hides a regression rather
than a prover limitation, so it should fail there until the allocation is
fixed.

Verified: `qa/allocations.jl` completes on Julia 1.10, the version that was
failing; `GROUP=QA` and `GROUP=Core` both pass on 1.12, which is what CI runs.

Stacked on #1119 - it carries the `_lu_from_cacheval` fix this needs.
